Refactor logic to make cleaner
